List any server which has both IPv4 and IPv6 addresses once.
This makes use of bit 2 of net_enabled (whether IPv6 is preferred).
People who set bit 1 (enabling use of IPv6) should also set this bit.

FIXME: This is a little different from how the server browsers handle this,
in that the game servers which have both types of address MUST have the same
grouping and be returned in the same packet. The reason for this is that
there's an annoying race condition somewhere which means that if servers
whose info was received in an earlier packet have their IP address info
replaced, there may be corruption of the server list (incorrect server names
etc.).