Audio: Port context creation and hide OpenAL in a single File.
Context creation is now made by Audio.cpp, snd_openal.cpp became empty
and was promptly removed. OpenAL headers are only included in ALObjects
(the name is a bit misleading it is not only the objects but also the
global OpenAL functions). To do that we assumed ALuint will always be
typedefed to unsigned.