Fix sub-frame input handling.
This needs IN_Frame() to be called often, maybe more than once per frame.
Timestamps reflect when this was called rather than the actual event times,
so we work around this by calling it from within the delay loop in
Com_Frame(). Wake-up within that loop is approx. every 50ms, so the
minimum calling frequency should be 20Hz.