Remove the GetGameState syscall, refactor gamestate
GameState was a char array in which string were allocated. This commit
replaces it with an array of std::strings. That way we can easily send
it to the VM. Also the VM doesn't load the whole game state again on a
"cs" server command but just replaces the right config string.